Refactor functions to AuthorizedSenderDomainController
[MAILPOET-5782]
