+++ mailpoet/lib/Util/Notices/SenderDomainAuthenticationNotices.php
@@ -2,7 +2,6 @@
 
 namespace MailPoet\Util\Notices;
 
-use MailPoet\Newsletter\Statistics\NewsletterStatisticsRepository;
 use MailPoet\Services\AuthorizedSenderDomainController;
 use MailPoet\Services\Bridge;
 use MailPoet\Settings\SettingsController;
@@ -10,15 +9,10 @@ use MailPoet\Util\FreeDomains;
 use MailPoet\Util\Helpers;
 use MailPoet\Util\License\Features\Subscribers;
 use MailPoet\WP\Notice;
-use MailPoetVendor\Carbon\Carbon;
 
 class SenderDomainAuthenticationNotices {
   const FREE_MAIL_KB_URL = 'https://kb.mailpoet.com/article/259-your-from-address-cannot-be-yahoo-com-gmail-com-outlook-com';
   const SPF_DKIM_DMARC_KB_URL = 'https://kb.mailpoet.com/article/295-spf-dkim-dmarc';
-
-  const ENFORCEMENT_START_TIME = '2024-02-01 00:00:00 UTC';
-
-  const INSTALLED_AFTER_NEW_RESTRICTIONS_OPTION = 'installed_after_new_domain_restrictions';
 
   private SettingsController $settingsController;
 
@@ -30,22 +24,18 @@ class SenderDomainAuthenticationNotices {
 
   private Bridge $bridge;
 
-  private NewsletterStatisticsRepository $newsletterStatisticsRepository;
-
   public function __construct(
     SettingsController $settingsController,
     Subscribers $subscribersFeatures,
     FreeDomains $freeDomains,
     AuthorizedSenderDomainController $authorizedEmailsController,
-    Bridge $bridge,
-    NewsletterStatisticsRepository $newsletterStatisticsRepository
+    Bridge $bridge
   ) {
     $this->settingsController = $settingsController;
     $this->subscribersFeatures = $subscribersFeatures;
     $this->freeDomains = $freeDomains;
     $this->authorizedSenderDomainController = $authorizedEmailsController;
     $this->bridge = $bridge;
-    $this->newsletterStatisticsRepository = $newsletterStatisticsRepository;
   }
 
   public function getDefaultFromAddress(): string {
@@ -60,34 +50,12 @@ class SenderDomainAuthenticationNotices {
     return $this->freeDomains->isEmailOnFreeDomain($this->getDefaultFromDomain());
   }
 
-  // TODO: Remove after the enforcement date has passed
-  public function isEnforcementOfNewRestrictionsInEffect(): bool {
-    return Carbon::now() >= Carbon::parse(self::ENFORCEMENT_START_TIME);
-  }
-
-  public function isNewUser(): bool {
-    $installedVersion = $this->settingsController->get('version');
-
-    // Setup wizard has not been completed
-    if ($installedVersion === null) {
-      return true;
-    }
-
-    $installedAfterNewDomainRestrictions = $this->settingsController->get(self::INSTALLED_AFTER_NEW_RESTRICTIONS_OPTION, false);
-
-    if ($installedAfterNewDomainRestrictions) {
-      return true;
-    }
-
-    return $this->newsletterStatisticsRepository->countBy([]) === 0;
-  }
-
   public function init($shouldDisplay): ?Notice {
     if (
       !$shouldDisplay
       || !$this->bridge->isMailpoetSendingServiceEnabled()
       || in_array($this->getDefaultFromDomain(), $this->authorizedSenderDomainController->getFullyVerifiedSenderDomains(true))
-      || $this->isNewUser()
+      || $this->authorizedSenderDomainController->isNewUser()
       || $this->isFreeMailUser() && $this->subscribersFeatures->getSubscribersCount() <= AuthorizedSenderDomainController::LOWER_LIMIT
     ) {
       return null;
@@ -114,7 +82,7 @@ class SenderDomainAuthenticationNotices {
   }
 
   public function isErrorStyle(): bool {
-    if (!$this->isEnforcementOfNewRestrictionsInEffect()) {
+    if (!$this->authorizedSenderDomainController->isEnforcementOfNewRestrictionsInEffect()) {
       return false;
     }
     if (
@@ -132,7 +100,7 @@ class SenderDomainAuthenticationNotices {
   }
 
   public function getNoticeContentForFreeMailUsers(int $contactCount): string {
-    if (!$this->isEnforcementOfNewRestrictionsInEffect()) {
+    if (!$this->authorizedSenderDomainController->isEnforcementOfNewRestrictionsInEffect()) {
       // translators: %1$s is the domain of the user's default from address, %2$s is a rewritten version of their default from address, %3$s is HTML for an 'update sender' button, and %4$s is HTML for a Learn More button
       return sprintf(__("<strong>Update your sender email address to a branded domain by February 1st, 2024 to continue sending your campaigns.</strong>
 <span>Starting on February 1st, 2024, MailPoet will no longer be able to send from email addresses on shared 3rd party domains like <strong>%1\$s</strong>. Please change your campaigns to send from an email address on your site's branded domain. Your emails will temporarily be sent from <strong>%2\$s</strong>.</span> <p>%3\$s &nbsp; %4\$s</p>", 'mailpoet'),
@@ -165,7 +133,7 @@ class SenderDomainAuthenticationNotices {
   }
 
   public function getNoticeContentForBrandedDomainUsers(bool $isPartiallyVerified, int $contactCount): string {
-    if (!$this->isEnforcementOfNewRestrictionsInEffect() || $isPartiallyVerified || $contactCount <= AuthorizedSenderDomainController::LOWER_LIMIT) {
+    if (!$this->authorizedSenderDomainController->isEnforcementOfNewRestrictionsInEffect() || $isPartiallyVerified || $contactCount <= AuthorizedSenderDomainController::LOWER_LIMIT) {
       // translators: %1$s is HTML for an 'authenticate domain' button, %2$s is HTML for a Learn More button
       return sprintf(__("<strong>Authenticate your sender domain to improve email delivery rates.</strong>
 <span>Major mailbox providers require you to authenticate your sender domain to confirm you sent the emails, and may place unauthenticated emails in the \"Spam\" folder. Please authenticate your sender domain to ensure your marketing campaigns are compliant and will reach your contacts.</span><p>%1\$s &nbsp; %2\$s</p>", 'mailpoet'),
