+++ mailpoet/tests/integration/Util/Notices/SenderDomainAuthenticationNoticesTest.php
@@ -4,9 +4,6 @@ namespace integration\Util\Notices;
 
 use MailPoet\Services\AuthorizedSenderDomainController;
 use MailPoet\Settings\SettingsController;
-use MailPoet\Test\DataFactories\Newsletter;
-use MailPoet\Test\DataFactories\StatisticsNewsletters;
-use MailPoet\Test\DataFactories\Subscriber;
 use MailPoet\Util\Notices\SenderDomainAuthenticationNotices;
 use MailPoetVendor\Carbon\Carbon;
 
@@ -161,35 +158,5 @@ class SenderDomainAuthenticationNoticesTest extends \MailPoetTest {
     $message = $this->notice->getNoticeContentForBrandedDomainUsers(true, 1200);
     $this->assertStringContainsString('Authenticate your sender domain to improve email delivery rates.', $message);
     $this->assertStringContainsString('Authenticate domain', $message);
-  }
-
-  public function testUserIsNewIfTheyHaveNotCompletedWelcomeWizard(): void {
-    $this->settings->delete('version');
-    $this->assertTrue($this->notice->isNewUser());
-  }
-
-  public function testUserIsNewIfTheyInstalledAfterNewRestrictionsImplemented(): void {
-    $this->settings->set('version', MAILPOET_VERSION);
-    $this->settings->set('installed_after_new_domain_restrictions', '1');
-    $this->assertTrue($this->notice->isNewUser());
-    // also true even if they've sent emails
-    (new StatisticsNewsletters((new Newsletter())->withSendingQueue()->create(), (new Subscriber())->create()))->create();
-    $this->assertTrue($this->notice->isNewUser());
-  }
-
-  public function testUserIsNotNewIfTheyDoNotHaveTheSettingAndHaveSentEmails(): void {
-    $this->settings->set('version', MAILPOET_VERSION);
-    $this->settings->delete('installed_after_new_domain_restrictions');
-    // No emails yet, so they are still "new"
-    $this->assertTrue($this->notice->isNewUser());
-    (new StatisticsNewsletters((new Newsletter())->withSendingQueue()->create(), (new Subscriber())->create()))->create();
-    $this->assertFalse($this->notice->isNewUser());
-  }
-
-  public function testItKnowsWhenNewRestrictionsStartGettingEnforced(): void {
-    Carbon::setTestNow(Carbon::parse('2024-01-31 00:00:00 UTC'));
-    $this->assertFalse($this->notice->isEnforcementOfNewRestrictionsInEffect());
-    Carbon::setTestNow(Carbon::parse('2024-02-01 00:00:01 UTC'));
-    $this->assertTrue($this->notice->isEnforcementOfNewRestrictionsInEffect());
   }
 }
